Board Folder Refactor
