+++ lab07/src/src/BoardManager/BoardGenerator/iBoardGenerator.java
@@ -1,7 +1,10 @@
-package BoardManager;
+package BoardManager.BoardGenerator;
+
+import BoardManager.Cells.iCell;
 
 public interface iBoardGenerator {
-    
+
     public iCell[][] generateNewBoard();
+
     public void notifyNewBoard();
 }
