+++ lab07/src/src/BoardManager/BoardLogic/BoardLogicController.java
@@ -1,4 +1,4 @@
-package BoardManager;
+package BoardManager.BoardLogic;
 
 import Utils.Position;
