+++ lab07/src/src/BoardManager/BoardLogic/PositionObservable.java
@@ -1,4 +1,4 @@
-package BoardManager;
+package BoardManager.BoardLogic;
 
 import Utils.BaseObervable;
 import Utils.Position;
