+++ lab07/src/src/BoardManager/BoardLogic/iBoardLogicController.java
@@ -1,4 +1,4 @@
-package BoardManager;
+package BoardManager.BoardLogic;
 
 public interface iBoardLogicController extends iMoveObservable {
