+++ lab07/src/src/BoardManager/BoardLogic/iMoveObservable.java
@@ -1,4 +1,4 @@
-package BoardManager;
+package BoardManager.BoardLogic;
 
 import Utils.Position;
 import Utils.iBaseObservable;
