+++ lab07/src/src/BoardManager/BoardState/iBoardStateController.java
@@ -1,4 +1,4 @@
-package BoardManager;
+package BoardManager.BoardState;
 
 import Utils.Position;
