+++ lab07/src/src/BoardManager/BoardView/PrintableBoardState.java
@@ -1,4 +1,4 @@
-package BoardManager;
+package BoardManager.BoardView;
 
 /**
  * PrintableBoardState
