+++ lab07/src/src/BoardManager/BoardView/iBoardView.java
@@ -1,4 +1,4 @@
-package BoardManager;
+package BoardManager.BoardView;
 
 public interface iBoardView {
     public PrintableBoardState exportPrintableBoardState();
