+++ lab07/src/src/BoardManager/Cells/EnemyCell.java
@@ -1,0 +1,1 @@
+package BoardManager.Cells;
