+++ lab07/src/src/BoardManager/Cells/iCell.java
@@ -1,4 +1,6 @@
-package BoardManager;
+package BoardManager.Cells;
+
+import BoardManager.BoardLogic.iBoardLogicController;
 
 public interface iCell {
     public void activateInteraction(iBoardLogicController controller);
