--- lab07/src/src/BoardManager/MoveObservable.java
@@ -1,7 +1,0 @@
-package BoardManager;
-
-import Utils.BaseObervable;
-import Utils.Position;
-
-public abstract class MoveObservable extends BaseObervable<iMoveObserver, Position> {
-}
