+++ lab07/src/src/BoardManager/iBoardStateManager.java
@@ -1,5 +1,8 @@
 package BoardManager;
 
+import BoardManager.BoardGenerator.iBoardGenerator;
+import BoardManager.BoardState.iBoardStateController;
+import BoardManager.BoardView.iBoardView;
 import HeroManager.iHeroManConsumer;
 import UIManager.iUIManConsumer;
